Updating Contributing section of README.md
Prior to this change, the Contributing section referenced cloning https://github.com/harikt/the-book-on-aura , the previous location for the Aura manual. Now that the manual has been merged into the repository, the Contributing section has been rewritten to remove references to the book on aura repository (using a modified version of the Composer README.md contributing section).