Fixes for issue #41 https://github.com/auraphp/auraphp.github.com/issues/41
Fixed the Macro examples as they were referencing the old v2 config
structure of config/default/config.file which has been replaced to just
config/config.file. Also put in references to the methods that need to
be modified with the example code so it was clear where in the file
changes needed to be made.